relaxation dtype of bert

+++ commons/model.py
@@ -64,8 +64,7 @@ def get_model_tokenizer(
         tokenizer = AutoTokenizer.from_pretrained(os.path.dirname(__file__).replace("commons","tokenizer"))
         model = AutoModelForCausalLM.from_pretrained(
             MODEL_KEY2IDS[model_key],
-            attn_implementation='eager',
-            torch_dtype=torch.bfloat16,
+            attn_implementation='eager'
         )
         model, tokenizer = adjust_tokenizer(model, tokenizer)
         return model, tokenizer, None
